Revert "20240624 작업"
