--- src/components/webzine/WebzineList.jsx
@@ -1,18 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-
-const WebzineList = ({ webzine, closeMenu }) => {
-  const navigate = useNavigate();
-  return (
-    <li
-      onClick={() => {
-        navigate(`/webzineDetail/${webzine._id}`);
-        closeMenu();
-      }}
-    >
-      {webzine.title}
-    </li>
-  );
-};
-
-export default WebzineList;
