+++ src/layouts/Header.jsx
@@ -23,13 +23,12 @@ const Header = () => {
     document.body.style.overflow = isMenuOpen ? 'auto' : 'hidden';
   };
 
-  const openProfileMenu = (e) => {
-    e.preventDefault();
+  const openProfileMenu = () => {
     setIsProfileMenuOpen(!isProfileMenuOpen);
+    document.body.style.overflow = isProfileMenuOpen ? 'auto' : 'hidden';
   };
 
-  const handleLogout = (e) => {
-    e.preventDefault();
+  const handleLogout = () => {
     localStorage.removeItem('userid'); // 로컬 스토리지에서 userid 제거
     localStorage.removeItem('token'); // 로컬 스토리지에서 토큰 제거
