+++ src/styles/Chatbot.module.css
@@ -1,132 +1,6 @@
 .chatbot {
   position: fixed;
   z-index: 9999999999999999;
-  right: 3rem;
-  bottom: 3rem;
-}
-.chatbot > button {
-  width: 100px;
-  height: 100px;
-  border-radius: 100%;
-  background-color: var(--main-color);
-  color: var(--white);
-  font-size: 2.3rem;
-  position: relative;
-  box-shadow: 0px 5px 15px rgba(0, 0, 0, 15%);
-}
-.chatbot > button:hover::before,
-.chatbot > button:hover::after {
-  opacity: 1;
-}
-.chatbot > button:hover::before {
-  transform: translate(-50%, -50%) scale(0);
-}
-.chatbot > button:hover::after {
-  transform: translate(-50%, -50%) scale(1.25);
-}
-.chatbot > button::before,
-.chatbot > button::after {
-  content: '';
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 100%;
-  height: 100%;
-  border-radius: 100%;
-  z-index: 20;
-  opacity: 0;
-  transition: 0.5s;
-}
-.chatbot > button::before {
-  border: 1px solid var(--main-color);
-  transform: translate(-50%, -50%) scale(1.5);
-  z-index: 20;
-}
-.chatbot > button::after {
-  background-color: var(--sub-color);
-  transform: translate(-50%, -50%) scale(0);
-  z-index: 10;
-}
-.chatbot > button span {
-  position: absolute;
-  top: -15%;
-  left: -190%;
-  background-color: var(--white);
-  color: var(--main-color);
-  font-size: 1rem;
-  border: 1px solid var(--main-color);
-  padding: 1rem;
-  border-radius: 30px;
-}
-.chatbot > button i {
-  z-index: 40;
-  position: relative;
-}
-
-.chatbot .chatContainer.show {
-  display: block;
-}
-.chatbot .chatContainer {
-  display: none;
-  z-index: 100;
-  position: absolute;
-  bottom: 0;
   right: 0;
-  width: 450px;
-  height: fit-content;
-  max-height: 650px;
-  border-radius: 50px;
-  background-color: #f0f0f0;
-  box-shadow: 5px 5px 15px rgba(0, 0, 0, 15%);
-  padding: 2rem;
-}
-.chatbot .chatContainer > div {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-.chatbot .chatContainer > div > div {
-  max-height: 468px;
-  overflow: hidden;
-  overflow-y: scroll;
-  display: flex;
-  flex-direction: column;
-  gap: 0.85rem;
-}
-.chatbot .chatContainer h2 {
-  margin-bottom: 1rem;
-  font-size: 1.5em;
-}
-.chatbot .chatContainer > div > div > p > strong {
-  display: none;
-}
-.chatbot .chatContainer > form {
-  display: flex;
-  flex-wrap: nowrap;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.8rem;
-  margin-top: 1rem;
-}
-.chatbot .chatContainer > form input {
-  flex: 1;
-  padding: 0.95rem;
-  border-radius: 100px;
-  border: 1px solid var(--black);
-}
-.chatbot .chatContainer > form button {
-  width: 50px;
-  height: 50px;
-  border-radius: 100%;
-  background-color: var(--main-color);
-  color: var(--white);
-}
-.chatbot .chatConClose {
-  position: absolute;
-  top: 27px;
-  right: 27px;
-  font-size: 27px;
-  font-weight: bold;
-  background-color: transparent;
+  bottom: 0;
 }
