+++ src/styles/Header.module.css
@@ -120,8 +120,8 @@
 }
 .gnb .logon .profileImg a {
   display: block;
-  width: 50px;
-  height: 50px;
+  width: 37px;
+  height: 37px;
   border-radius: 100%;
   background-color: #d9d9d9;
 }
@@ -131,7 +131,7 @@
 }
 .gnb .logon .toggleMenu.profile {
   position: absolute;
-  top: 60px;
+  top: 50px;
   right: 0;
   display: flex;
   flex-direction: column;
@@ -142,5 +142,5 @@
   border-radius: 5px;
   padding: 1.5rem 2.5rem;
   white-space: nowrap;
-  box-shadow: 0 5px 10px rgb(0 0 0 / 10%);
+  box-shadow: 0 5px 10px rgb(215 215 215 / 70%);
 }
