+++ src/styles/myRecipe/MyRecipeDetail.module.css
@@ -1,8 +1,10 @@
 .mySwiper {
   margin-bottom: 40px;
 }
-.swiperSlide .image {
-  width: 100%;
+.swiperSlide {
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .title,
