+++ src/styles/myRecipe/MyRecipeList.module.css
@@ -19,11 +19,6 @@
   gap: 1rem;
 }
 
-.gridContainer > li {
-  overflow: hidden;
-  border-radius: 5px;
-}
-
 .actionBar {
   display: flex;
   align-items: center;
