+++ src/styles/style.css
@@ -227,23 +227,3 @@ main section .btnMore i {
 .mainFeed .imgArea {
   grid-template-columns: repeat(auto-fit, minmax(33%, 1fr));
 }
-
-/* chatbot */
-.chatbot .user,
-.chatbot .bot {
-  width: fit-content;
-  max-width: 90%;
-  padding: 0.75rem 1rem;
-  line-height: 1.25;
-}
-.chatbot .user {
-  align-self: flex-end;
-  background-color: var(--white);
-  border-radius: 20px 20px 5px 20px;
-}
-.chatbot .bot {
-  align-self: flex-start;
-  background-color: var(--main-color);
-  color: var(--white);
-  border-radius: 5px 20px 20px 20px;
-}
