+++ src/styles/webzine/Webzine.module.css
@@ -3,9 +3,10 @@
   top: 0;
   left: 0;
   width: 100%;
-  min-height: 100vh;
+  min-height: 100%;
   z-index: 9999999;
   background-color: var(--white);
+  overflow-y: scroll;
 }
 
 #webzine header {
@@ -111,7 +112,7 @@
   width: 100%;
   min-height: fit-content;
 }
-#webzine .listArea > ul li {
+#webzine .listArea > ul li a {
   display: inline-block;
   width: 60%;
   padding-bottom: 2rem;
@@ -121,9 +122,8 @@
   font-family: 'pretendard-bold';
   font-size: 2.188em;
   transition: 0.5s;
-  cursor: pointer;
 }
-#webzine .listArea > ul li:hover {
+#webzine .listArea > ul li a:hover {
   color: var(--main-color);
 }
 
@@ -196,9 +196,6 @@
 }
 #webzine.dark .contArea > div:nth-of-type(1) > button {
   background-color: var(--white);
-  color: var(--black);
-}
-#webzine.dark .contArea > div:nth-of-type(3) a {
   color: var(--black);
 }
